refactor(prompts): split catalog by domain
