+++ Sources/JarvisCore/Prompts/JarvisPrompts.swift
@@ -1,315 +1,7 @@
-import Foundation
-
-/// The single audit surface for predefined text sent to AI models.
+/// Namespace for predefined text sent to AI models.
 ///
-/// Keep transport payloads, tool schemas, and dynamic user/session content with their owning
-/// subsystems. Every app-owned instruction, description, wrapper, and observation those payloads
-/// send to a model belongs here.
-public enum JarvisPrompts {
-    public enum Coach {
-        public static let system = """
-        # Identity
-        You are Jarvis, a calm, sharp technical-interview coach for behavioral, system-design, and coding
-        interviews. Help without interrupting productive thinking.
-
-        # Context
-        - "me:" is the user you coach. "them:" is the interviewer or caller. Speak only to "me"; never
-          answer "them" directly.
-        - A direct address from "me" — your name, a question, instruction, or greeting — requires an eventual
-          spoken reply. "them:" is context; offer "me" a tip only when useful.
-        - New speech appears under "New since last turn" with [mm:ss] timestamps. A
-          "(no speech for ...)" marker means quiet, not a request. Longer quiet makes being stuck more likely,
-          but does not prove it.
-        - You can see the screen only through capture_screen. A fresh screenshot or OCR in the current input
-          counts as current screen context.
-        - OCR text is a reading aid that garbles the odd token; the screenshot image is ground truth. Before
-          asserting a specific line or token is wrong, verify it in the image — if you can only see it in
-          OCR, frame the tip as something to double-check ("verify line 18 uses ==") rather than as a defect.
-
-        # Action policy
-        Choose exactly one action on each model response, in this priority order:
-
-        1. Direct address from "me": bypass the fragment gate. If a specific, correct reply depends on
-           missing current visible information, continue to the screen gate below. Otherwise call speak.
-        2. Fragment gate: when a non-silence request contains new speech, call stay_silent only if all of
-           it is incomplete or likely mistranscribed. Help/stuck signals and other meaningful speech bypass
-           this gate. If a reply is required despite uncertain transcription, hedge rather than correct it.
-        3. Screen gate: before speaking, capture when a specific, correct response depends on current visible
-           information that is absent from the conversation and no fresh capture result is available for this
-           request. This includes an explicit request to look or an unresolved reference to the current
-           question, code, error, diagram, document, or notes (for example, "this problem", "here", "my code",
-           or "one pass" without the problem). Never guess missing content. This gate applies to either speaker.
-           If "me" asked, call capture_screen now, then speak after the result. If only "them" spoke and no tip
-           is warranted, call stay_silent without capturing.
-        4. "me" is making steady progress: call stay_silent.
-        5. Progress is unclear, especially after silence: call capture_screen unless a fresh result is already
-           available. Then speak only if the user seems stuck; otherwise call stay_silent.
-        6. "me" is stuck: call speak with the next concrete step. Build on earlier tips instead of repeating
-           them.
-
-        A fresh capture result satisfies the screen gate for that request. Use it; do not capture again for
-        the same request.
-
-        # Tip style
-        Lead with the most useful point. Be brief, concrete, encouraging, and easy to read under pressure.
-        Prefer one pointed question or next step. Give a full solution only when "me" explicitly asks for it.
-        """
-
-        enum ToolDescription {
-            static let captureScreen = "Capture a fresh screenshot and OCR of visible interview "
-                + "context. Use when the next useful response depends on current screen information "
-                + "not already available; one fresh result satisfies that request."
-            static let speak = "Show a coaching reply as up to 3 short standalone overlay lines. "
-                + "Use one idea per line, aim under 12 words, and keep code on one line. Call only "
-                + "when a reply or tip is useful."
-            static let staySilent = "End this turn without speaking. Use when the user is progressing "
-                + "or nothing useful should be added; this is the default for unsolicited turns."
-        }
-
-        // Keep this a neutral marker. An earlier instruction to recapture, repeated in user-role
-        // history, biased the coach toward capturing on every quiet turn.
-        static let earlierImageStub = "[an earlier screenshot was here — no longer available]"
-        static let recognizedTextHeader =
-            "Text recognized on the captured window (on-device OCR — may contain "
-            + "errors; the screenshot image is ground truth):"
-        static let supersededRecognizedTextStub =
-            "[an earlier screen's OCR text was here — superseded by a newer capture]"
-        static let manualHintCaptureFailed =
-            "The screen capture requested for the manual hint failed."
-        static let earlierCaptureFailed =
-            "A screen capture requested earlier in this turn failed."
-        static let captureFailed = "screenshot failed"
-        static let captureSucceeded = "screenshot captured"
-        static let tipShown = "shown to the user"
-
-        static func newSpeech(_ text: String) -> String {
-            "New since last turn:\n\(text)"
-        }
-
-        static func silenceTrigger(timestamp: String, duration: String) -> String {
-            "[\(timestamp)] (no speech for \(duration))"
-        }
-
-        static func manualHintTrigger(timestamp: String) -> String {
-            "[\(timestamp)] The user pressed the hint shortcut. They want your single most useful "
-                + "hint about what's on their screen right now — answer using the attached screenshot "
-                + "and the recent transcript."
-        }
-
-        static func recognizedText(_ text: String) -> String {
-            "\(recognizedTextHeader)\n\(text)"
-        }
-
-        static func captureResult(recognizedText text: String?) -> String {
-            guard let text else { return captureSucceeded }
-            return "\(captureSucceeded)\n\n\(recognizedText(text))"
-        }
-
-        static func condensedHistory(_ summary: String) -> String {
-            "[session so far, condensed — earlier turns were summarized]\n\(summary)"
-        }
-    }
-
-    public enum HistorySummary {
-        public static let system = """
-        You condense a live coding-interview coaching session's history into a briefing the coach will \
-        rely on for the rest of the session. Keep, in this order: the interview problem statement (all \
-        load-bearing details); the user's current approach and how far they've got; every tip the coach \
-        already gave (so it isn't repeated); any open questions or requirements from the interviewer. \
-        Plain text, under 250 words. Output only the briefing.
-        """
-
-        static func input(_ messages: [ChatMessage]) -> String {
-            messages.map { message in
-                if let calls = message.toolCalls {
-                    return calls.map { "coach called \($0.name)" }.joined(separator: "\n")
-                }
-                if message.imageBase64JPEG != nil { return "[screenshot]" }
-                let text = message.text ?? ""
-                return message.role == .tool ? "tool result: \(text)" : text
-            }.joined(separator: "\n")
-        }
-    }
-
-    enum LocalAgent {
-        static let codexDirectResponse = """
-            Answer this decision request immediately without inspecting files, running commands,
-            browsing, planning, delegating, or invoking any Codex built-in tool. The capture_screen,
-            speak, and stay_silent names below are an output JSON protocol, not callable Codex tools.
-            """
-
-        static func roleBlock(_ role: String, text: String) -> String {
-            "[\(role)]\n\(text)"
-        }
-
-        static func assistantToolCall(name: String, argumentsJSON: String) -> String {
-            roleBlock("assistant", text: "{\"tool\":\"\(name)\",\"arguments\":\(argumentsJSON)}")
-        }
-
-        static func conversationHeading(isFirstTurn: Bool) -> String {
-            isFirstTurn ? "## Conversation" : "## New input"
-        }
-
-        static let screenshotPlaceholder = roleBlock("user", text: "(screenshot below)")
-
-        static func answerTrailer(forcedToolName: String?) -> String {
-            var trailer = "Answer now, following the tool protocol."
-            if let forcedToolName {
-                trailer += " You MUST call the `\(forcedToolName)` tool this turn."
-            }
-            return trailer
-        }
-
-        static func toolProtocol(tools: [ToolDef], toolChoice: ToolChoice) -> String {
-            var lines = [
-                "## Tool protocol",
-                "",
-                "You are the decision engine inside an automated harness — your reply is parsed "
-                    + "by a program, not read by a person. These are your tools:",
-            ]
-            for tool in tools {
-                lines.append("- \(tool.name) — \(tool.description)")
-                lines.append("  arguments JSON Schema: \(tool.parametersJSON)")
-            }
-            lines.append("")
-            lines.append(
-                "End your reply with a single line containing ONLY this JSON object (no code "
-                    + "fence, nothing after it): {\"tool\":\"<tool name>\",\"arguments\":{…}}. "
-                    + "Use {} for a tool with no arguments."
-            )
-            switch toolChoice {
-            case .required, .force:
-                // `.force` stays byte-identical to `.required` here so one forced hint does not
-                // rewrite the cacheable system prefix. Its tool name belongs in the turn trailer.
-                lines.append(
-                    "You MUST pick exactly one tool this turn — the JSON object is your entire answer."
-                )
-            case .auto:
-                lines.append("If no tool fits, reply with plain text instead of the JSON object.")
-            }
-            return lines.joined(separator: "\n")
-        }
-    }
-
-    public enum Transcription {
-        public static func context(for speaker: Speaker) -> String {
-            switch speaker {
-            case .me:
-                "A live technical-interview conversation captured from the local user's microphone. "
-                    + "This stream contains the user's speech and may include names, numbers, and "
-                    + "technical terminology."
-            case .them:
-                "A live technical-interview conversation captured from Mac system audio. This "
-                    + "stream contains other participants' speech and may include names, numbers, "
-                    + "and technical terminology."
-            }
-        }
-    }
-
-    enum Evaluation {
-        static func sessionAudit(
-            sessionDirectoryPath: String,
-            transcriptFilename: String,
-            trafficFilename: String,
-            activityFilename: String,
-            reportFilename: String
-        ) -> String {
-            """
-            You are auditing one session of Jarvis, a proactive macOS coaching assistant: it transcribes \
-            the user's (and their interlocutor's) speech, optionally screenshots the user's screen, and an \
-            LLM decides each turn to speak a short overlay tip, look at the screen, or stay silent. The \
-            session's memory is client-managed: each request is [instructions] + history + new turn, built \
-            append-only so the provider's prompt cache keeps hitting; old history is periodically compacted \
-            into a summary by a separate "summarizer" client.
-
-            YOUR WORKSPACE is a checkout of the Jarvis source repository. The audited session lives at:
-                \(sessionDirectoryPath)
-            with these inputs in that directory:
-              - `\(transcriptFilename)` — the session's wire-level LLM traffic. It OPENS with a \
-            "=== deterministic metrics ===" table computed from the raw traffic (per-call and \
-            aggregate input / cache-read / cache-write / output tokens and cost, plus a per-model \
-            breakdown). Trust its known numbers and availability labels: `—` means unavailable, not zero, \
-            and a `known (N unavailable)` value is partial, not a session total. Quote only what the table \
-            supports; never recompute a total by eye. The rest is \
-            one block per API call, in order — your PRIMARY narrative input. To keep it compact, content \
-            byte-identical to the previous same-tag call is elided and marked "(unchanged)" — those \
-            markers are exactly where the prompt cache SHOULD be hitting.
-              - `\(trafficFilename)` — the same traffic un-elided. Because the transcript elides \
-            byte-identical repeats, ANY cardinal count (stub occurrences, OCR dumps, marker repetitions) \
-            MUST be counted here, not from the elided transcript.
-              - `\(activityFilename)` — the COMPLETE sanitized human-facing coaching record: heard \
-            speech, manual hints, every brain action, and every fixed stop/degrade notice, with stable \
-            event kinds in `k`. Read the file itself in full; it is deliberately NOT filtered, summarized, \
-            or copied into `\(transcriptFilename)`. Use it whenever you need the user-visible sequence or \
-            lifecycle consequence. Treat `session ended by error` as a session-level UX failure, but not \
-            an end caused by the user, app quit, or a new session; distinguish all of them from a \
-            recoverable `listening continues` notice. A single call \
-            timeout that stopped the whole session is catastrophic even if the wire-level failure itself \
-            looks ordinary.
-
-            Provider records can appear together — read the right fields for each:
-              - OpenAI Responses: `response.usage` with `input_tokens`, \
-            `input_tokens_details.cached_tokens` (automatic prefix-cache hit), optional \
-            `cache_write_tokens`, and `output_tokens`; a truncated run shows `status:"incomplete"`. No \
-            per-call dollar cost. OpenAI `input_tokens` includes cached input, so mixed-provider token \
-            totals are kept separate.
-              - Claude Code warm query: `response.cli` with `total_cost_usd`, a call-level `usage` \
-            carrying Anthropic's `cache_creation_input_tokens` / `cache_read_input_tokens` split, and a \
-            `modelUsage` map (per-model usage + cost, including internal sidecar models like a haiku \
-            pass). One coaching attempt leases one preinitialized query; a capture follow-up sends only \
-            incremental input over that same query. Verify the turn boundary in \
-            `Sources/JarvisCore/Brain/Adapters/LocalAgent/ClaudeCode/ClaudeCodeRuntime.swift`.
-              - Codex app-server: `response.runtime` carries completed-turn metadata but currently no \
-            token, cache, or cost usage. Those cells must stay unavailable; never interpret them as zero. \
-            One coaching attempt owns one fresh ephemeral thread and sends capture follow-ups \
-            incrementally; verify `CodexAppServerRuntime.swift`.
-              - `shot-N.jpg` — the screenshots the model actually saw (base64 was redacted from the traffic).
-              - a prior `\(reportFilename)`, if this session was audited before.
-
-            Because you have the repo, DO NOT GUESS how the harness works — read it and verify. Load-bearing \
-            files: `Sources/JarvisCore/Prompts/JarvisPrompts.swift` (all predefined model-facing prompt \
-            copy), `Sources/JarvisCore/Coach/ToolDefs.swift` (tool schemas), \
-            `Sources/JarvisCore/Coach/CoachHistory.swift` (client-managed memory: append-only prefix, \
-            screenshot stubbing, compaction), `Sources/JarvisCore/Coach/CoachDriver.swift` (the event loop, \
-            the filler substance gate, compaction trigger), \
-            `Sources/JarvisCore/Brain/ReasoningEffort.swift` (`max_output_tokens` is a COMBINED \
-            reasoning+output budget — a tight cap recreates the documented `status:"incomplete"` \
-            truncation), and `Sources/JarvisCore/Config/Config.swift`. Before recommending a mechanism, \
-            grep for it: several (screenshot stubbing, bulky-tool-result compaction, filler gating, \
-            prompt_cache_key pinning) already exist.
-
-            Write a markdown report with exactly these sections:
-
-            ## Context engineering
-            Inefficiencies in what the harness sends: cache-busting prefix changes (input marked changed \
-            that should have been stable), redundant or stale content re-billed every call, oversized \
-            instructions or tool schemas relative to their value, history-compaction timing/quality, \
-            screenshots kept or stubbed at the wrong time. Be quantitative with the usage numbers, and \
-            name the code that produces each behavior.
-
-            ## Issues and errors
-            Transport errors, non-2xx responses, truncated runs (status=incomplete), tool-loop anomalies \
-            (repeated capture_screen, exhausted loops), and any contradiction between the code/instructions \
-            and the model's behavior.
-
-            ## Coaching quality
-            Given the transcript deltas the model saw: tips that were wrong, late, redundant, or noisy; \
-            silences that should have been tips and vice versa.
-
-            ## Recommendations
-            Concrete changes, ordered by expected impact, each tied to evidence above. Prefix EACH with a \
-            label:
-              - `[confirmed]` — you verified it against the code in this checkout (cite the file/symbol). \
-            Use this ONLY when you actually read the relevant code; "confirmed" means confirmed against the \
-            implementation, not merely consistent with the traffic.
-              - `[hypothesis]` — plausible from the traffic but not (or not yet) checked against the code.
-
-            Cite call numbers (call #N) as evidence throughout. If the data doesn't support a finding, say \
-            so rather than speculating. Before finalizing, re-check every number in your draft against the \
-            metrics table and against `\(trafficFilename)`, and correct or delete any that disagree. Output \
-            ONLY the markdown report — it is saved as `\(reportFilename)` in the session directory (the \
-            harness prepends a one-line provenance stamp, so don't add your own).
-            """
-        }
-    }
-}
+/// The files in this directory are the single audit surface. Keep transport payloads, tool schemas,
+/// and dynamic user/session content with their owning subsystems. Every app-owned instruction,
+/// description, wrapper, and observation those payloads send to a model belongs in the matching
+/// domain extension here.
+public enum JarvisPrompts {}
