+++ Tests/JarvisCoreTests/Diagnostics/AgenticEvaluationTests.swift
@@ -52,8 +52,9 @@ import Foundation
         #expect(prompt.contains("## Recommendations"))
         #expect(prompt.contains("[confirmed]"))
         #expect(prompt.contains("[hypothesis]"))
-        #expect(prompt.contains("Prompts/JarvisPrompts.swift"))
+        #expect(prompt.contains("Sources/JarvisCore/Prompts/"))
         #expect(prompt.contains("all predefined model-facing prompt"))
+        #expect(prompt.contains("split by domain"))
         #expect(prompt.contains("CoachHistory.swift"))
         #expect(prompt.contains(AgenticEvaluation.reportFilename))
         #expect(prompt.contains(ActivityLog.filename))
