+++ scripts/eval-session.sh
@@ -4,7 +4,7 @@
 # The sole evaluator runs through an agentic CLI (Claude Code / Codex) whose workspace is this repo
 # checkout PLUS the complete session directory. The auditor reads the full Activity log and raw
 # traffic whenever useful, then verifies each finding against the harness's own code
-# (JarvisPrompts.swift, CoachHistory.swift, CoachDriver.swift, ToolDefs.swift, …) instead of guessing
+# (Prompts/, CoachHistory.swift, CoachDriver.swift, ToolDefs.swift, …) instead of guessing
 # from a reduced prompt.
 #
 # The script and Activity's Evaluate button both call the same Foundation-only evaluator. The agent
